Support any kind of buteo-contacts accounts
Do not hardcode Google as the only supported Online Accounts provider;
instead, enumerate the available account providers and present them in
the UI.

Note that for some weird reason the address-book-app.application file
for Online Accounts is not part of this project, but resides in the
buteo-sync-plugins-contacts project. Eventually, this needs to be fixed.