Fix the freeze when removing a contact
For some reason, the second parameter of PopupUtils.open() makes a big
difference here: without it, we get the

    QQmlExpression: Attempted to evaluate an expression in an invalid context

warning printed on the logs, and the application stops responding. Take
this chance to make some refactoring of the dialog and page code, since
the dialog should not have any knowledge about which its parent windows
are.