ContactListModel: allow deletion of multiple aggregate contacts
Removal of an aggregate contact and its constituents was only working
for one contact at a time, since the relationship model's "participant"
property would be immediately overwritten when setting the next contact
to fetch (the previous lookup would not be completed, since lookups are
asynchronous).

To fix that, we instantiate the relationship model on the heap, so that
every call to enumerateContactIds() would use its own model; in this way
we allow parallel execution of the method.